Push test environment classes to the package level
