+++ server/src/test/java/io/spine/server/event/given/bus/GivenEventMessage.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2019, TeamDev. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.server.event.given.bus;
+
+import io.spine.test.event.ProjectCreated;
+import io.spine.test.event.ProjectId;
+import io.spine.test.event.ProjectStarted;
+
+public class GivenEventMessage {
+
+    private static final ProjectStarted PROJECT_STARTED = projectStarted(EventBusTestEnv.PROJECT_ID);
+
+    private GivenEventMessage() {
+    }
+
+    static ProjectStarted projectStarted() {
+        return PROJECT_STARTED;
+    }
+
+    static ProjectCreated projectCreated(ProjectId id) {
+        return ProjectCreated.newBuilder()
+                             .setProjectId(id)
+                             .build();
+    }
+
+    static ProjectStarted projectStarted(ProjectId id) {
+        return ProjectStarted.newBuilder()
+                             .setProjectId(id)
+                             .build();
+    }
+}
